moved makeToken into security folder
