+++ User/loginRouter.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const User = require('./User');
-const { makeToken } = require('./auth');
-const { authenticate } = require('./security');
+// const { makeToken } = require('./auth');
+const { authenticate, makeToken } = require('./security');
 
 router.post('/', authenticate, (req, res) => {
   // const { username, password } = req.body;
